requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
cryptography==42.0.8
defusedxml==0.7.1
distlib==0.3.8
dj-database-url==2.2.0
dj-rest-auth==2.1.11
Django==4.2
django-allauth==0.54.0
django-cors-headers==3.10.1
django-crispy-forms==2.1
django-email==0.1.10
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.14.0
filelock==3.14.0
idna==3.7
Markdown==3.6
MouseInfo==0.1.3
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
pillow==10.3.0
pipenv==2023.12.1
platformdirs==4.2.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyJWT==2.8.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python3-openid==3.2.0
pytweening==1.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==69.5.1
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.2
